Tests/LibWeb/Text/expected/css/animated-inherited-font-size-line-height.txt

parent font-size: 20px
child font-size: 20px
child line-height: 40px
